agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "strands-agents-builder>=0.1.10",
    "strands-agents-tools>=0.2.18",
    "strands-agents[gemini,openai]>=1.20.0",
    "pydantic-settings>=2.0.0",
]
